+++ test/UnitTyping.cpp
@@ -13,18 +13,17 @@
 
 #include <config.h>
 
-#include <random>
-#include <iostream>
-
+#include <COOLWSD.hpp>
 #include <Exceptions.hpp>
 #include <Log.hpp>
 #include <Unit.hpp>
 #include <UnitHTTP.hpp>
+#include <WebSocketSession.hpp>
 #include <helpers.hpp>
-#include <COOLWSD.hpp>
-
 #include <wsd/TileDesc.hpp>
-#include <net/WebSocketSession.hpp>
+
+#include <random>
+#include <iostream>
 
 using namespace ::helpers;
