+++ test/helpers.hpp
@@ -11,8 +11,16 @@
 
 #pragma once
 
-#include <test/testlog.hpp>
 #include <test/lokassert.hpp>
+#include <test/testlog.hpp>
+
+#include <Socket.hpp>
+#include <Common.hpp>
+#include <WebSocketSession.hpp>
+#include <common/ConfigUtil.hpp>
+#include <common/Util.hpp>
+#include <tools/COOLWebSocket.hpp>
+#include <wsd/TileDesc.hpp>
 
 #include <Poco/BinaryReader.h>
 #include <Poco/Dynamic/Var.h>
@@ -25,15 +33,6 @@
 #include <Poco/Net/NetException.h>
 #include <Poco/Path.h>
 #include <Poco/URI.h>
-
-#include <Common.hpp>
-#include "Socket.hpp"
-#include "common/FileUtil.hpp"
-#include <tools/COOLWebSocket.hpp>
-#include <common/ConfigUtil.hpp>
-#include <common/Util.hpp>
-#include <net/WebSocketSession.hpp>
-#include <wsd/TileDesc.hpp>
 
 #include <chrono>
 #include <filesystem>
